requirements.txt
astrapy>=0.7.1
python-dotenv>=1.0.0
opensearch-py>=2.4.1
tqdm>=4.66.1
faker>=22.6.0 